requirements.txt
py_setenv
wget
GitPython
tqdm